Temporary fix in parsing of ids of length 50
The bug happens with sequences with ids of lenght 50. If the
fasta headers have <50 characters, IMGT keeps them. If
they are >50 characters long, IMGT truncates to 49 and replaces
some charaters with '_'. If they are 50, they are only trucanted,
no character replacement is applied. MakeDB imgt was expecting
that truncation and replacement would be applied headers >49
characters long. Because the replacement was not being applied
to sequences with headers 50 characters long, MakeDb was not
able to match correctly the fasta file and the
imgt files, and header annotation was not transferred to the db
files.

The current change works with recent results of IMGT/HighV-quest
(version 1.8.3). But I am not sure it works correctly when
parsing old imgt files. Probably it doesn't. I don't have old results
with sequences with lengths 49, 50 and 51 that I can use to test, and
see what modifications where being applied. I sent an email to IMGT,
and I am waiting for a response.